% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{all_pathways}
\alias{all_pathways}
\title{All BioCarta pathways}
\usage{
all_pathways()
}
\value{
A vector of pathway IDs (the primary pathway IDs on BioCarta).
}
\description{
All BioCarta pathways
}
\details{
The original BioCarta website (biocarta.com) is retired, but the full list of pathways can be found from archived websites such as
\url{https://web.archive.org/web/20170122225118/https://cgap.nci.nih.gov/Pathways/BioCarta_Pathways} or \url{https://www.gsea-msigdb.org/gsea/msigdb/human/genesets.jsp?collection=CP:BIOCARTA}.
}
\examples{
all_pathways()
}
