% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{genes_in_pathway}
\alias{genes_in_pathway}
\title{Genes in a pathway}
\usage{
genes_in_pathway(pathway)
}
\arguments{
\item{pathway}{A BioCarta pathway ID, a MSigDB ID or a \code{biocarta_pathway} object.}
}
\value{
A character vector of Entrez IDs.
}
\description{
Genes in a pathway
}
\examples{
genes_in_pathway("h_RELAPathway")
}
