% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{get_pathway}
\alias{get_pathway}
\title{Get a single pathway}
\usage{
get_pathway(pathway_id)
}
\arguments{
\item{pathway_id}{A biocarta pathway ID.}
}
\value{
A \code{biocarta_pathway} object.
}
\description{
Get a single pathway
}
\examples{
get_pathway("h_RELAPathway")
}
