% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{get_pathway_image}
\alias{get_pathway_image}
\alias{image_dimension}
\title{Download the pathway image}
\usage{
get_pathway_image(pathway)

image_dimension(pathway)
}
\arguments{
\item{pathway}{A pathway ID or a \code{biocarta_pathway} object.}
}
\value{
A \code{raster} object.
}
\description{
Download the pathway image
}
\examples{
img = get_pathway_image("h_RELAPathway")
class(img)

image_dimension("h_RELAPathway")
}
