% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{makeContext.biocarta_pathway_grob}
\alias{makeContext.biocarta_pathway_grob}
\alias{grobWidth.biocarta_pathway_grob}
\alias{grobHeight.biocarta_pathway_grob}
\title{Internal functions for drawing the pathway grob}
\usage{
\method{makeContext}{biocarta_pathway_grob}(x)

\method{grobWidth}{biocarta_pathway_grob}(x)

\method{grobHeight}{biocarta_pathway_grob}(x)
}
\arguments{
\item{x}{A \code{grob} returned by \code{\link[=pathwayGrob]{pathwayGrob()}}.}
}
\value{
A \code{grob} object.

A \code{unit} object.

A \code{unit} object.
}
\description{
Internal functions for drawing the pathway grob
}
