% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{grid.pathway}
\alias{grid.pathway}
\alias{pathwayGrob}
\title{Draw a BioCarta pathway}
\usage{
grid.pathway(
  pathway,
  color = NULL,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = NULL,
  height = NULL,
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  default.units = "npc",
  name = NULL
)

pathwayGrob(
  pathway,
  color = NULL,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = NULL,
  height = NULL,
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  default.units = "npc",
  name = NULL
)
}
\arguments{
\item{pathway}{A pathway ID or a `biocarta_pathway` object.}

\item{name}{}
}
\description{
Draw a BioCarta pathway
}
