% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BIOCARTA_PATHWAYS}
\alias{BIOCARTA_PATHWAYS}
\alias{PATHWAY2BC}
\alias{PATHWAY2ENTREZ}
\alias{BC2ENTREZ}
\title{Pre-computated datasets}
\format{
An object of class \code{list} of length 314.

An object of class \code{data.frame} with 4428 rows and 2 columns.

An object of class \code{data.frame} with 5196 rows and 2 columns.

An object of class \code{data.frame} with 1739 rows and 2 columns.
}
\usage{
BIOCARTA_PATHWAYS

PATHWAY2BC

PATHWAY2ENTREZ

BC2ENTREZ
}
\description{
Pre-computated datasets
}
\details{
\code{BIOCARTA_PATHWAYS}: A list of pathways.
\code{PATHWAY2BC}: A two-column data frame.
\code{PATHWAY2ENTREZ}: A two-column data frame.
\code{BC2ENTREZ}: A two-column data frame.

\code{BCID} is an internal ID type in BioCarta. It is more like an ID for the nodes in the pathways.
}
\keyword{datasets}
