% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{get_pathway_image}
\alias{get_pathway_image}
\title{Download the pathway image}
\usage{
get_pathway_image(pathway_id)
}
\arguments{
\item{pathway_id}{A pathway ID.}
}
\value{
A `raster` object.
}
\description{
Download the pathway image
}
\examples{
img = get_pathway_image("h_RELAPathway")
class(img)
}
