% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocarta.R
\name{print.biocarta_pathway}
\alias{print.biocarta_pathway}
\title{Print the biocarta_pathway object}
\usage{
\method{print}{biocarta_pathway}(x, ...)
}
\arguments{
\item{x}{A \code{biocarta_pathway} object.}

\item{...}{Other arguments}
}
\value{
Nothing.
}
\description{
Print the biocarta_pathway object
}
\examples{
p = get_pathway("h_RELAPathway")
p
}
